# docs/reference/api-spec.yaml
...
openapi: 3.1.0
info:
  title: SceneFlow Backend
  version: 0.1.0
paths:
...
  /api/projects/{project_id}/assets:
    get:
      tags:
      - assets
      summary: List Assets
      operationId: list_assets_api_projects__project_id__assets_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Assets Api Projects  Project Id  Assets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - assets
      summary: Create Asset
      operationId: create_asset_api_projects__project_id__assets_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Asset Api Projects  Project Id  Assets Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/assets/{asset_id}:
    patch:
      tags:
      - assets
      summary: Update Asset
      operationId: update_asset_api_projects__project_id__assets__asset_id__patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Asset Api Projects  Project Id  Assets  Asset Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - assets
      summary: Delete Asset
      operationId: delete_asset_api_projects__project_id__assets__asset_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}/assets/merge:
    post:
      tags:
      - assets
      summary: Merge Assets
      operationId: merge_assets_api_projects__project_id__assets_merge_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeAssetsRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Merge Assets Api Projects  Project Id  Assets Merge Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/projects:
    get:
      tags:
      - projects
      summary: List Projects
      description: 'Every series, each carrying its current episode''s shots and a summary of the rest.


        Batched rather than serialized one project at a time: a list of series would otherwise

        run two queries per row. A project with no episode yet reads as empty here — creating

        one is a write, and a GET has no business doing that.'
      operationId: list_projects_api_projects_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Projects Api Projects Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_projects_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Project Api Projects Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Generation Job
      description: 'One job, for a client waiting on the work it queued.


        Its own endpoint rather than filtering the project list: a panel waiting on one image

        should not pull every job in the series on a three-second interval.'
      operationId: get_generation_job_api_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Generation Job Api Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/jobs/{job_id}/cancel:
    post:
      tags:
      - jobs
      summary: Cancel Generation Job
      operationId: cancel_generation_job_api_jobs__job_id__cancel_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Cancel Generation Job Api Jobs  Job Id  Cancel Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/jobs/{job_id}/retry:
    post:
      tags:
      - jobs
      summary: Retry Generation Job
      operationId: retry_generation_job_api_jobs__job_id__retry_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Retry Generation Job Api Jobs  Job Id  Retry Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateAssetRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        kind:
          type: string
          enum:
          - image
          - video
          - audio
          title: Kind
        data:
          type: string
          maxLength: 80000000
          minLength: 20
          title: Data
      additionalProperties: false
      type: object
      required:
      - name
      - kind
      - data
      title: CreateAssetRequest
...
    CreateProjectRequest:
      properties:
        title:
          type: string
          maxLength: 80
          title: Title
          default: ''
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        coverPrompt:
          type: string
          maxLength: 4000
          title: Coverprompt
          default: ''
        originalScript:
          type: string
          maxLength: 200000
          title: Originalscript
          default: ''
        productionSettings:
          $ref: '#/components/schemas/ProductionSettingsRequest'
      additionalProperties: false
      type: object
      title: CreateProjectRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MergeAssetsRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 4000
          title: Description
          default: ''
        kind:
          type: string
          enum:
          - image
          - video
          - audio
          title: Kind
        assetIds:
          items:
            type: string
          type: array
          maxItems: 64
          minItems: 1
          title: Assetids
      additionalProperties: false
      type: object
      required:
      - name
      - kind
      - assetIds
      title: MergeAssetsRequest
...
    ProductionSettingsRequest:
      properties:
        mode:
          anyOf:
          - type: string
            enum:
            - comic
            - drama
          - type: 'null'
          title: Mode
        aspectRatio:
          anyOf:
          - type: string
            enum:
            - '9:16'
            - '16:9'
            - '1:1'
          - type: 'null'
          title: Aspectratio
        width:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 256.0
          - type: 'null'
          title: Width
        height:
          anyOf:
          - type: integer
            maximum: 4096.0
            minimum: 256.0
          - type: 'null'
          title: Height
        fps:
          anyOf:
          - type: integer
            enum:
            - 24
            - 30
          - type: 'null'
          title: Fps
        targetDurationMs:
          anyOf:
          - type: integer
            maximum: 600000.0
            minimum: 10000.0
          - type: 'null'
          title: Targetdurationms
        language:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Language
        stylePrompt:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Styleprompt
        negativePrompt:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Negativeprompt
        currentStage:
          anyOf:
          - type: string
            enum:
            - script
            - bible
            - storyboard
            - audio
            - timeline
            - export
          - type: 'null'
          title: Currentstage
      additionalProperties: false
      type: object
      title: ProductionSettingsRequest
      description: 'Every field optional so a PATCH can carry just the one the user changed.


        Range checks live here for the OpenAPI schema; `project_service.production_settings`

        stays the single place that resolves defaults and cross-field consistency.'
...
    UpdateAssetRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 4000
          - type: 'null'
          title: Description
        data:
          anyOf:
          - type: string
            maxLength: 80000000
            minLength: 1
          - type: 'null'
          title: Data
      additionalProperties: false
      type: object
      title: UpdateAssetRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError